Revert "fix: improve Kimi VL model robustness and MoE parallelizer defaults"
This reverts commit a185e34e06322ae3b52a80767b57b216f3656a57.

+++ nemo_automodel/components/moe/parallelizer.py
@@ -184,7 +184,7 @@ def apply_fsdp(
     if mp_policy is None:
         mp_policy = MixedPrecisionPolicy(
             param_dtype=torch.bfloat16,
-            reduce_dtype=torch.bfloat16,
+            reduce_dtype=torch.float32,
             output_dtype=torch.bfloat16,
             cast_forward_inputs=True,
         )
